Move adapter, mappers to Business Layer
