--- src/db/mongo/schemas/create-category.schema.ts
@@ -1,3 +1,0 @@
-export class CreateCategorySchema {
-    displayName: string;
-}
