--- src/db/mongo/schemas/create-product.schema.ts
@@ -1,5 +1,0 @@
-export class CreateProductSchema {
-    displayName: string;
-    category: string;
-    price: number;
-}
